+++ releasenotes.md
@@ -212,6 +212,7 @@
 - Sometimes generating introspection info would not be in the global scope causing a crash #1586.
 - @tag on macros cannot be retrieved with tagof #1582
 - Taking the $typeof of a wildcard optional returns `void!`.
+- Fix bug with enums with jump tables #1840.
 
 ### Stdlib changes
 - Remove unintended print of `char[]` as String